requirements.txt
pandas==2.2.1
numpy
yfinance==0.2.36
alpaca-trade-api==3.2.0
python-dotenv==1.0.1
requests==2.31.0
streamlit
plotly
ccxt
google-generativeai
Pillow
yahooquery
finvizfinance
pycoingecko
cryptocompare
beautifulsoup4
newspaper3k
ntscraper
customtkinter
